chore: Move service_outofstoremessage.go to pkg/outofstoremessage
